Fix Bag Labels shrinking, again
